Improvements for SMTP logging (#40)
* config.py: Encode settings as str objects

json.load() in Python2 imports strings as unicode strings. If SMTP
logging is used and the SMTP server supports CRAM MD5, these unicode
strings are processed by Python's hmac module which only works properly
with str objects. Hence, we have to encode the settings as str objects.

* docs/alerts/email.rst: Add the secure flag to doc

If the SMTP server supports STARTTLS or TLS, we can set "secure" in the
logging config to enable it. This is mandatory for some SMTP servers.