Updated config.py to be Python 3 compatible
Changed from .itteritems() to just .items() That's a little less efficient for
Python 2 because items() loads the full list into memory rather than creating
an iterator and getting one at a time but for the size of the lists we are
using it doesn't matter.

Also only use 'unicode' if we are using Python 2.