Changed from hpfeeds to hpfeeds3
hpfeeds3 is a python 3 compatible fork of hpfeeds that seems to be
actively maintained.

At the time of this writing the last commit on hpfeeds was May 22, 2015
~4 years ago while the last commit on hpfeeds3 was Nov 23, 2018 ~6 months
ago.

As far as I can tell from my limited testing hpfeeds3 is backwards compatible
with python 2 as well, although if it causes issues we could install
different versions like we do with wsgiref.

https://github.com/Jc2k/hpfeeds3