Fixed a bug in the mysql implementation
The mysql implementation has a bug where 2 hashtags instead of 1 were added to the error string. In addition to
this, the error code for a denied access (SQL_STATE_ACCESS_DENIED) is 1 zero too short. This makes that a hashtag
is visible in the error message returned to the client which it shouldn't be.

According to the documentation: https://dev.mysql.com/doc/internals/en/packet-ERR_Packet.html
if CLIENT_PROTOCOL_41 is among the capabilities, the state requires 1 hashtag and a string of
exactly 5 characters.

This leaves attackers able to detect the honeypot in 2 ways:

1. Attempt to login and notice the hashtag in the error message, this however triggers a warning on the honeypot,
so one detects the honeypot but is still found out.

2. Send a packet out of order, triggering the SQL_STATE_PKT_ORDER error. This doesn't trigger a warning as the
connection is dropped and no login attempt is made. However it's still possible to notice the 2 hashtags returning
in the error message.