Fix inconsistent SMB Logging
Previously the srcHost was dstHost and srcHost was blank. Many other logged fields were empty or transposed. Any access to the SMB honeypot would on occasion blow up because the number of variables the results of the result of splitting on '|' was assigned into didn't match the number of array items resulting from the split.

The regex is simpler and is used solely to identify lines of interest in the audit file.
Now we use an split to throw away the log metadata, with a second split on '|' which is assigned to a new object.
We then use the new object to populate a set of distinctly named variables. This might not be the cleanest or most efficient way of doing it, but it does show the assumptions made at the time this code was written and I think also makes it easier to maintain when the log format changes.

I've also removed the Samba 3 support because surely no one uses it?