Update setup.py
Changes allow for package discovery without relying on git metadata. This removes a dependency on git when installing opencanary and simplifies installation from git archives which lack metadata.
These changes are noted to break the automated CI tests on Ubuntu 18.04. This is caused by the tests using setuptools 39.0.1-2 which does not have find_namespace_packages and is below minimum the requirements.